backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-google-genai
langchain-chroma
chromadb
pydantic
requests
langchain-huggingface
sentence-transformers